logic_server/requirements.txt
# app
uvicorn
pydantic
email-validator
fastapi-users[sqlalchemy]
# jwt
python-jose[cryptography]
# db
psycopg2
databases[postgresql]
SQLAlchemy
alembic